requirements.txt
Kivy==2.1.0
kivymd==1.1.1
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
pycryptodomex==3.17
pyperclip==1.8.2
python-dotenv==0.21.1
seaborn==0.11.2
SQLAlchemy==1.4.41
